requirements.txt
flask==3.0.3
flask-sqlalchemy>=3.0.0
openai>=1.68.0
markdown>=3.7
flask-migrate==4.1.0
quill-delta[html]==1.0.3